Исправление ошибок тестового запуска

See merge request hilkovich/album_stories!2

+++ bot/models/users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, DateTime
+from sqlalchemy import Column, Integer, DateTime, BigInteger
 from sqlalchemy.orm import declarative_base
 
 from database.connection import Engine
@@ -9,7 +9,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = "users"
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True)
+    telegram_id = Column(BigInteger, unique=True)
     created_on = Column(DateTime)
